Fix redirect_to issue
Since the user is directed over to Google for authentication, url
parameters are lost, along with WordPress' `redirect_to` which sends
the user to where they were trying to go before they were forced to
log in.

This fixes the issue by passing the query parameters over to
Google as `state` which then get's sent back.